+++ packages/plugin-autocapture-browser/src/constants.ts
@@ -43,3 +43,4 @@ export const AMPLITUDE_VISUAL_TAGGING_HIGHLIGHT_CLASS = 'amp-visual-tagging-sele
 export const DATA_AMP_MASK_ATTRIBUTES = 'data-amp-mask-attributes';
 
 export const MAX_MASK_TEXT_PATTERNS = 25;
+export const MAX_ATTRIBUTE_LENGTH = 128;