+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.27](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.26...@amplitude/analytics-node@1.5.27) (2025-12-09)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
 ## [1.5.26](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.25...@amplitude/analytics-node@1.5.26) (2025-11-21)
 
 **Note:** Version bump only for package @amplitude/analytics-node