+++ packages/segment-session-replay-plugin/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.0.0-beta.51](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.50...@amplitude/segment-session-replay-plugin@0.0.0-beta.51) (2025-12-09)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
+# [0.0.0-beta.50](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.49...@amplitude/segment-session-replay-plugin@0.0.0-beta.50) (2025-12-05)
+
+**Note:** Version bump only for package @amplitude/segment-session-replay-plugin
+
+
+
+
+
 # [0.0.0-beta.49](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/segment-session-replay-plugin@0.0.0-beta.48...@amplitude/segment-session-replay-plugin@0.0.0-beta.49) (2025-11-21)
 
 **Note:** Version bump only for package @amplitude/segment-session-replay-plugin